VAE: fix beta scaling
Pyro's TraceELBO uses a TraceMessenger to capture the trace, which
does so in its _pyro_post_sample function. That messenger is at the
very bottom of the stack, so its _pyro_post_sample function runs
first, therefore the trace does not capture any modifications to the
message made after that.

Apply the scale in _pyro_sample instead