implement a translator based on conditional normalizing flows
This uses a coupling neural spline flow with a rotation matrix
after each block. The spline flow only transforms data with a certain
range (by default [-5, 5]). To make sure that the input falls within
this range, ideas from batch normalization are borrowed: During the
first training epoch, mean and variance of the cNF input are accumulated.
These are then used to standardize the cNF input.

This version translates betweent the deterministic latent embeddings
(mean and standard deviation of the variational distribution).