robustify sampling INN translator
The previous version used the empirical variance of the latent mean
to scale samples from the variational distribution. This turned out
to not be sufficient: Some dimensions can have very small variance
of the mean, but a large variance of the variational distribution,
such that samples from the variational distribution have a much larger
variance than assumed by the scaling. This led to highly inflated
values after scaling, which were outside the effective of the neural
spline flow.

This version assumes that the empirical variance of the variational mean
and the variational variance are independent, such that they can be
added up to yield the variance of a sample from the variational distribution.