implement prediction for INN translators
For the latent translator, parameters of the destination variational
distribution are sampled from the cNF. Using these parameters, a sample
from the variational posterior of the observed variables is drawn.

For the sampling translator, a sample from the source variational
distribution is drawn, This sample is used to condition the cNF, from
which a sample is drawn. This sample is decoded by the destination VAE
and a sample of the variational posterior of the observed variables
is returned.

Both translators currently return a sample of observed data, including
sequencing depth scaling. This means that currently, observed data from
the destination is still required.

Both translators take 1000 samples for each batch. To avoid having to
store all samples in memory, streaming statistics of the sampling distribution
are calculated. In particular, these are mean, variance, and several approximate
quantiles. The statistics per minibatch are immediately saved to disk
using Zarr.