implement double-sampling INN translator
This maps samples from the source variational distribution to parameters
of the target variational distribution. During prediction, a sample
is drawn from the source variational distribution. A sample is then
drawn from the cNF to get parameters from the target variational
distribution. Finally, a sample from the target variational distribution
is drawn to get the prediction